Update style for consistency with image_classification example
